Bower version mismatch
Bower detects and reports a difference between the semver git tag and the version in this bower.json file. One approach would be to ensure that it's updated for each release - but that's a pain. So a better approach is simply to remove it and allow the semver git tag to be used without contradiction.